services/image-processor/requirements.txt
cohere==4.47
pymongo==4.6.1
redis==5.0.1
Pillow==10.2.0
numpy==1.26.3
wordcloud==1.9.3
